google map working

+++ BlazorWASM-AppLogic/Domain/Models/Coordinate.cs
@@ -2,10 +2,10 @@
 
 public class Coordinate
 {
-    public long latitude{ get; set; }
-    public long longitude{ get; set; }
+    public float latitude{ get; set; }
+    public float longitude{ get; set; }
 
-    public Coordinate(long latitude, long longitude)
+    public Coordinate(float latitude, float longitude)
     {
         this.latitude = latitude;
         this.longitude = longitude;
